Keep JSP work within the source-size gate (Fixes #522)
CI source-size failed because three files exceeded the 1000-line hard limit.
Fixed by extraction rather than by relaxing the gate.

jsp_host becomes a module directory with the HTTP request parsing and response
writing split into a wire submodule that holds no protocol state.

manager gains a jsp_launch module owning instrumentation and revocation, and a
manager_liveness module holding the read-only liveness and signature queries,
following the existing manager_passthrough precedent. manager.rs was already
983 lines before this branch, so it had no headroom for a new feature.

events.rs sat exactly at the limit, so its two observation variants are
consolidated into one Observation variant carrying an ObservationEvent enum
defined in a new state module. That groups the related lifecycle events and
moves their payload types out of the capped file.
