+++ src/messages/event_conversion.rs
@@ -5,6 +5,7 @@
 //! @pseudocode component-004 lines 46-50
 
 use crate::state::AppEvent;
+use crate::state::observation_events::ObservationEvent;
 
 use super::{
     ActionsMessage, AppMessage, ErrorsMessage, IssuesMessage, ModalMessage, PersistenceMessage,
@@ -197,10 +198,14 @@ impl AppMessage {
             AppEvent::AgentStatusChanged(id, status) => {
                 Self::Runtime(RuntimeMessage::AgentStatusChanged(id, status))
             }
-            AppEvent::ObservationUpdated(id, generation, observation) => Self::Runtime(
-                RuntimeMessage::ObservationUpdated(id, generation, observation),
-            ),
-            AppEvent::ObservationCleared(id, generation) => {
+            AppEvent::Observation(ObservationEvent::Updated(id, generation, observation)) => {
+                Self::Runtime(RuntimeMessage::ObservationUpdated(
+                    id,
+                    generation,
+                    observation,
+                ))
+            }
+            AppEvent::Observation(ObservationEvent::Cleared(id, generation)) => {
                 Self::Runtime(RuntimeMessage::ObservationCleared(id, generation))
             }
             AppEvent::PersistenceLoadSuccess => Self::Persistence(PersistenceMessage::LoadSuccess),
@@ -718,10 +723,10 @@ impl From<RuntimeMessage> for AppEvent {
             RuntimeMessage::RestartAgent(id) => Self::RestartAgent(id),
             RuntimeMessage::AgentStatusChanged(id, status) => Self::AgentStatusChanged(id, status),
             RuntimeMessage::ObservationUpdated(id, generation, observation) => {
-                Self::ObservationUpdated(id, generation, observation)
+                Self::Observation(ObservationEvent::Updated(id, generation, observation))
             }
             RuntimeMessage::ObservationCleared(id, generation) => {
-                Self::ObservationCleared(id, generation)
+                Self::Observation(ObservationEvent::Cleared(id, generation))
             }
         }
     }
